requirements.txt
astroid==2.3.3
autopep8==1.4.4
Click==7.0
Flask==1.1.1
gunicorn==20.0.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
pyserial==3.4
six==1.13.0
Werkzeug==0.16.0
wrapt==1.11.2